requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
stripe==11.4.1
blinker==1.9.0
click==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
